+++ examples/09_mint_native_token.rs
@@ -39,7 +39,7 @@ async fn main() -> Result<()> {
 
     // Wait for transaction to get included
     account
-        .retry_until_included(&transaction.block_id.expect("no block created yet"), None, None)
+        .retry_transaction_until_included(&transaction.transaction_id, None, None)
         .await?;
 
     account.sync(None).await?;