+++ CHANGELOG.md
@@ -26,6 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `RequiredStorageDeposit` and `RequiredStorageDepositDto` types;
 - `account/types/balance` module;
 - `AccountBuilder::with_bech32_hrp()`;
+- `Account::retry_transaction_until_included()`;
+- `RetryTransactionUntilIncluded` to message interface account methods;
 - `AccountMethod::RequestFundsFromFaucet` to message interface;
 
 ### Changed