+++ bindings/java/CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Throw `InitializeWalletException` on failed initialization;
 - Wallet.java `clearListeners`, `destroy` and `initLogger` methods;
 - Event listening (`WalletEvent` and `TransactionProgressEvent`);
+- `AccountHandle::retryTransactionUntilIncluded`;
 
 ### Changed